seperated drag code and fixed lint errors
